+++ src/app/cart/cart.module.ts
@@ -1,18 +1,23 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { CartComponent, CartItemComponent } from './components';
+import { CartComponent, CartItemComponent, CartListComponent } from './components';
+
+import { SharedModule } from '../shared';
 
 @NgModule({
   declarations: [
    CartComponent,
-   CartItemComponent
+   CartItemComponent,
+   CartListComponent
   ],
   exports: [
     CartComponent
   ],
   imports: [
-    CommonModule
+    CommonModule,
+
+    SharedModule
   ]
 })
 export class CartModule { }