+++ src/app/cart/components/index.ts
@@ -1,2 +1,3 @@
 export * from './cart';
 export * from './cart-item';
+export * from './cart-list';