+++ src/app/shared/directives/index.ts
@@ -1,0 +1,5 @@
+import { HighlightDirective } from './highlight.directive';
+
+export const directives = [
+  HighlightDirective
+];