+++ changelog.md
@@ -1,3 +1,12 @@
+### Home Task 2
+
+- Changed app prefix to 'sh';
+- Add OnPush strategy to presentation components;
+- Add possibility to change count of product in CartItemComponent;
+- Add calculating of totalAmount and totalCount to CartService;
+- Add title text using ViewChild and Renderer2;
+- Add directive [hightlight] for changing background of hovered cart-item. 
+
 ### Home Task 1
 
 - Created FirstComponent with simple markup and added to AppComponent